requirements.txt
fastapi[standard]
langchain
langchain_core
langchain_openai
langgraph
openai
pydantic
streamlit
typing_extensions
websocket-client
websockets
